Quick fixes (#40)
* Update Partfloor Email

Update the party floor confirmation email to include important information.

* CM Name lookup

Add a way for merchants to be looked up with their legal name. Fixes #36

* Fix Email Lockouts

Removed Uniqueness constraint for emails, and changed the check to make sure it matches the current event and the email before telling an applicant that the email has already applied. Fixes #38

* Registration Reminder Button

Added a button to send an email to a merchant to remind them to register for the con.

* Fix Competitor Count

Fix Competitor Count on dashboard.

* Update Requirements

Update the requirements.